LayoutTests/fast/custom-elements/CustomElementRegistry-expected.txt


PASS CustomElementRegistry interface must have define as a method 
PASS customElements.define must throw with an invalid name 
PASS customElements.define must throw when there is already a custom element of the same name 
PASS customElements.define must throw when there is already a custom element with the same class 
PASS customElements.define must throw when the element interface is not a constructor 
PASS customElements.define must get "prototype" property of the constructor 
PASS customElements.define must rethrow an exception thrown while getting "prototype" property of the constructor 
PASS customElements.define must throw when "prototype" property of the constructor is not an object 
PASS customElements.define must get callbacks of the constructor prototype 
PASS customElements.define must rethrow an exception thrown while getting callbacks on the constructor prototype 
PASS customElements.define must rethrow an exception thrown while converting a callback value to Function callback type 
PASS customElements.define must get "observedAttributes" property on the constructor prototype when "attributeChangedCallback" is present 
PASS customElements.define must rethrow an exception thrown while getting observedAttributes on the constructor prototype 
PASS customElements.define must rethrow an exception thrown while converting the value of observedAttributes to sequence<DOMString> 
PASS customElements.define must rethrow an exception thrown while iterating over observedAttributes to sequence<DOMString> 
PASS customElements.define must rethrow an exception thrown while retrieving Symbol.iterator on observedAttributes 
PASS customElements.define must not throw even if "observedAttributes" fails to convert if "attributeChangedCallback" is not defined 
PASS customElements.define must define an instantiatable custom element 
